requirements.txt
openai>=1.40.0
pyyaml>=6.0
rich>=13.0
python-dotenv>=1.0
